+++ FoodOrderApp/FoodOrderApp/Models/PaymentType.cs
@@ -14,7 +14,7 @@ namespace FoodOrderApp.Models
 
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public long PaymentTypeId { get; set; }
+        public int PaymentTypeId { get; set; }
         public string PaymentName { get; set; }
 
         public virtual ICollection<Orders> Orders { get; set; }
